requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
djangorestframework-gis==1.1
drf-yasg==1.21.7